NodeSplit Axis calc
Started work on the Node Split. Added functionality to find the Axis with the biggest spread, so later on i can Split the node along that Axis.

Next Steps: Sort Content along axis, find Pivot entry, setup two lists for each of the extremes and partition the entries to each side of the pivot.

Then do the actual Split logic.