Code Cleanup
Added comments, cleaned the Code and while doing so found a bug with object positions in teh tree. Added a separate dictionary of tree positions, so remove etc. dont look for the object on its changed/ new position but rather at the position where it is in the tree. Sadly that brought back the null Ref on Leaf search.